Create day-25-Running Time and Complexity
I've learnt that if a number is divisible by another number less or equal to the square root of the first number... it is NOT prime. That is the reason for O(sqrt(n)) run time.

I'll stick to this method when it comes to determining whether a number is prime. I've had a little trouble with using Fermat's Little Theorem right now because some composite numbers can sneak pass the test.